- Add global exception handling with custom exceptions for `NotFoundException` and `BadRequestException` - Implement CORS configuration for flexible API access - Refactor update methods in `UserService` and `ProjectService` to include `id` as a parameter and improve exception handling - Add validation constraints and default values to `User` and `Project` entities - Update `UserController` with CRUD endpoints and validations - Fix database URL property in `application.properties` for correct PostgreSQL connection string

+++ src/main/java/com/crudactivity/projectsync/service/UserServiceImpl.java
@@ -2,14 +2,18 @@ package com.crudactivity.projectsync.service;
 
 
 import com.crudactivity.projectsync.entity.User;
+import com.crudactivity.projectsync.exception.NotFoundException;
 import com.crudactivity.projectsync.repository.UserRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
-public class UserServiceImpl implements GenericService<User> {
+public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
 
@@ -24,6 +28,7 @@ public class UserServiceImpl implements GenericService<User> {
 
     @Override
     public Optional<User> getById(Long id) {
+        User user = userRepository.findById(id).orElseThrow(() -> new NotFoundException("User with id " + id + "Not Found"));
         return userRepository.findById(id);
     }
 
@@ -34,16 +39,21 @@ public class UserServiceImpl implements GenericService<User> {
 
     @Override
     public void deleteById(Long id) {
+        if(getById(id).isEmpty()){
+            throw new NotFoundException("User with Id " + id + " Not Found");
+        }
         userRepository.deleteById(id);
     }
 
 
     @Override
-    public User update(User user) {
-        Optional<User> optionalUser = userRepository.findById(user.getId());
+    public User update(Long id, User user) {
+        Optional<User> optionalUser = userRepository.findById(id);
         if(optionalUser.isEmpty()){
-            throw new RuntimeException();
+            throw new NotFoundException("User with Id " + user.getId() + " Not Found");
         }
+        user.setId(id);
+        user.setCreateAt(optionalUser.get().getCreateAt());
         return userRepository.save(user);
     }
 }
